requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
python-multipart==0.0.9
websockets==12.0
httpx==0.27.2
motor==3.6.0
redis==5.0.8
aioredis==2.0.1
numpy<2.0
faiss-cpu==1.8.0.post1

sentence-transformers==3.0.1
Pillow==10.4.0
pytesseract==0.3.13
structlog==24.4.0
tenacity==9.0.0
orjson==3.10.7
